plaintext evaluate
