+++ src/implement/service_http/evaluate.rs
@@ -1,7 +1,6 @@
 use bytes::Bytes;
 use http_body::Body;
 use http_body_util::{BodyExt, Collected};
-use regex::Regex;
 use serde::{Deserialize, Serialize};
 
 #[cfg(feature = "json")]
@@ -14,6 +13,7 @@ use crate::{
     },
     service::{
         destinations::Destinations,
+        evaluate::plaintext::PlaintextEvaluate,
         evaluator::{Evaluator, RequestResult},
     },
 };
@@ -55,27 +55,6 @@ pub enum BodyEvaluate {
     Plaintext(PlaintextEvaluate),
     #[cfg(feature = "json")]
     Json(JsonEvaluate),
-}
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize, Default)]
-#[serde(deny_unknown_fields, rename_all = "kebab-case")]
-pub struct PlaintextEvaluate {
-    pub regex: Option<EvaluateTo<String>>,
-}
-impl PlaintextEvaluate {
-    pub fn plaintext_acceptable(&self, body: &Destinations<Bytes>, msg: &mut Vec<EvaluateError>) -> bool {
-        let _ = msg; // TODO body[d] can be failed
-        match &self.regex {
-            Some(EvaluateTo::All(regex)) => body.iter().all(|(_, b)| {
-                Regex::new(regex).map(|re| re.is_match(String::from_utf8_lossy(b).as_ref())).unwrap_or(false)
-            }),
-            Some(EvaluateTo::Destinations(dest)) => dest.iter().all(|(d, regex)| {
-                Regex::new(regex)
-                    .map(|re| re.is_match(String::from_utf8_lossy(body[d].as_ref()).as_ref()))
-                    .unwrap_or(false)
-            }),
-            None => true,
-        }
-    }
 }
 
 impl Coalesce for HttpResponse {
