+++ src/service/evaluate/mod.rs
@@ -1,2 +1,3 @@
 #[cfg(feature = "json")]
 pub mod json;
+pub mod plaintext;
